@@ -1,53 +1,16 @@
-import React, { useState } from "react";
-import "./styles.css";
-import SignInForm from "./SignIn";
-import SignUpForm from "./SignUp";
+import { Routes, Route, BrowserRouter } from "react-router-dom"
+import SignInSignUp from "./screens/SignInSignUp";
+import Home from "./screens/Home";
+import LogedIn from "./screens/LogedIn";
 
 export default function App() {
-  const [type, setType] = useState("signIn");
-  const handleOnClick = (text) => {
-    if (text !== type) {
-      setType(text);
-      return;
-    }
-  };
-  const containerClass =
-    "container " + (type === "signUp" ? "right-panel-active" : "");
   return (
-    <div className="App">
-      <h1 className="project-title">Project UATU</h1>
-      <div className={containerClass} id="container">
-        <SignUpForm />
-        <SignInForm />
-        <div className="overlay-container">
-          <div className="overlay">
-            <div className="overlay-panel overlay-left">
-              <h1>Welcome Back!</h1>
-              <p>
-                To keep connected with us please login with your personal info
-              </p>
-              <button
-                className="ghost"
-                id="signIn"
-                onClick={() => handleOnClick("signIn")}
-              >
-                Sign In
-              </button>
-            </div>
-            <div className="overlay-panel overlay-right">
-              <h1>Hello, Friend!</h1>
-              <p>Enter your personal details and start journey with us</p>
-              <button
-                className="ghost "
-                id="signUp"
-                onClick={() => handleOnClick("signUp")}
-              >
-                Sign Up
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home />} exact />
+        <Route path="/login" element={<SignInSignUp />} exact />
+        <Route path="/app" element={<LogedIn />} exact />
+      </Routes>
+    </BrowserRouter>
   );
 }
